修改了数据持久化存储

+++ hdgl.db.store/src/hdgl/db/store/impl/hdfs/mapreduce/GraphInputStream.java
@@ -169,8 +169,8 @@ public class GraphInputStream extends InputStream{
 		}
 		if (bs[0] >= 128)
 		{
-			long tmp = (bs[0] << 24) + (bs[1] << 16) + (bs[2] << 8) + bs[3];
-			ret = (int) (tmp - (1 << 32));
+			long tmp = (((long)bs[0]) << 24) + (((long)bs[1]) << 16) + (((long)bs[2]) << 8) + (long)bs[3];
+			ret = (int) (tmp - (((long)1) << 32));
 		}
 		else
 		{
